requirements.txt
APScheduler==3.6.3
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
emoji==0.6.0
future==0.18.2
importlib-metadata==3.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.1
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
youtube-dl==2020.12.29
zipp==3.4.0
redis~=3.5.3